+++ backend/apid/routers/authentication.go
@@ -143,8 +143,6 @@ func (a *AuthenticationRouter) logout(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
-
-	return
 }
 
 // token handles logic for issuing new access tokens