+++ build.sh
@@ -38,7 +38,7 @@ install_deps () {
     go get gopkg.in/alecthomas/gometalinter.v1
     go get github.com/gordonklaus/ineffassign
     go get github.com/jgautheron/goconst/cmd/goconst
-    go get github.com/kisielk/errcheck
+    go get honnef.co/go/tools/cmd/megacheck
     go get github.com/golang/lint/golint
     go get github.com/UnnoTed/fileb0x
     install_golang_dep
@@ -74,7 +74,7 @@ build_tool_binary () {
 
     local outfile="target/${goos}-${goarch}/${subdir}/${cmd}"
 
-    GOOS=$goos GOARCH=$goarch go build -i -o $outfile ${REPO_PATH}/${subdir}/${cmd}/...
+    GOOS=$goos GOARCH=$goarch go build -o $outfile ${REPO_PATH}/${subdir}/${cmd}/...
 
     echo $outfile
 }
@@ -98,7 +98,7 @@ build_binary () {
     local ldflags+=" -X $version_pkg.BuildDate=${build_date}"
     local ldflags+=" -X $version_pkg.BuildSHA=${build_sha}"
 
-    CGO_ENABLED=0 GOOS=$goos GOARCH=$goarch go build -ldflags "${ldflags}" -i -o $outfile ${REPO_PATH}/${cmd}/cmd/...
+    CGO_ENABLED=0 GOOS=$goos GOARCH=$goarch go build -ldflags "${ldflags}" -o $outfile ${REPO_PATH}/${cmd}/cmd/...
 
     echo $outfile
 }
@@ -160,7 +160,7 @@ linter_commands () {
         exit 1
     fi
 
-    errcheck $(go list ./... | grep -v dashboardd | grep -v agent/assetmanager | grep -v scripts)
+    megacheck $(go list ./... | grep -v dashboardd | grep -v agent/assetmanager | grep -v scripts)
     if [ $? -ne 0 ]; then
         echo "Linting failed..."
         exit 1