+++ CHANGELOG.md
@@ -21,7 +21,8 @@ set when importing legacy settings.
 - Add event metrics type, implementing the Sensu Metrics Format.
 - Agents now try to reconnect to the backend if the connection is lost.
 - Added non-functional selections for resolving and silencing to web ui
-- Add LastOk to check type. This will be updated to reflect the last timestamp of a succesful check 
+- Add LastOk to check type. This will be updated to reflect the last timestamp
+of a successful check.
 
 ### Changed
 - Refactor Check data structure to not depend on CheckConfig. This is a breaking
@@ -39,6 +40,8 @@ except that HookConfig is now embedded in Hook.
 subcommand.
 - Agent backend URLs without a port specified will now default to port 8081.
 - Travis encrypted variables have been updated to work with travis-ci.org
+- Upgraded all builds to use Go 1.10.
+- Use megacheck instead of errcheck.
 
 ### Fixed
 - Fixed a bug in time.InWindow that in some cases would cause subdued checks to
@@ -49,6 +52,7 @@ characters.
 erroneously expire.
 - Resolved a bug in how an executor processes checks. If a check contains proxy
 requests, the check should not duplicately execute after the proxy requests.
+- Removed an erroneous validation statement in check handler.
 
 ## [2.0.0-alpha.17] - 2018-02-13
 ### Added